filter gem paths early
Bundler switches to debug output when the environment variable "DEBUG" is set. Due to the fact that this is a very generic name and used by different extensions and libraries, the output of "bundle list --paths" is more verbose than expected leading to the 'gem.list' error thrown by the Gem constructor when the line doesn't start with a forward slash.